extract ADC code into it's own driver
